Remove invalid sample
This file was committed by mistake. That project takes quite a few synchronous design decisions which would be hard to implement in webcs. Also a lot of additional code would need to be added to handle key/mouse input events (among other things) which only be used in this specific sample. So the idea was quickly scrapped.